// src/modules/cloud-storage/enums/media-type.enum.ts
export enum MediaTypeEnum{
    PAN = "PAN",
    AADHAR = "AADHAR",
    PROFILE_PHOTO = "PROFILE_PHOTO",
    SERVICE_IMAGE = "SERVICE_IMAGE",
    OUTLET_BANNER = "OUTLET_BANNER",
    SERVICE_VIDEO = "SERVICE_VIDEO",
    OUTLET_VIDEO = "OUTLET_VIDEO",
    OUTLET_GST = "OUTLET_GST",
    OUTLET_REGISTRATION = "OUTLET_REGISTRATION",
    OUTLET_MOU = "OUTLET_MOU"
}